Feature/optimized wpml handling

+++ app/PersonioIntegration/Position.php
@@ -260,8 +260,8 @@ class Position {
 			// assign the position to its terms.
 			foreach ( Taxonomies::get_instance()->get_taxonomies() as $taxonomy_name => $taxonomy ) {
 				if ( ! empty( $taxonomy['attr']['rewrite']['slug'] ) ) {
-					// first remove all existing relations if list is appended.
-					if ( $taxonomy['append'] ) {
+					// first remove all existing relations if list is appended (but not for language).
+					if ( $taxonomy['append'] && $taxonomy_name !== WP_PERSONIO_INTEGRATION_TAXONOMY_LANGUAGES ) {
 						wp_delete_object_term_relationships( $this->get_id(), $taxonomy_name );
 					}
 
@@ -524,11 +524,14 @@ class Position {
 	}
 
 	/**
-	 * Get the language of this object.
+	 * Return the language of this object.
 	 *
 	 * @return string
 	 */
 	public function get_lang(): string {
+		if( empty( $this->data['personioLanguages'] ) ) {
+			return Languages::get_instance()->get_main_language();
+		}
 		return $this->data['personioLanguages'];
 	}
